+++ deprecated/network/sample/src/main/java/ru/surfstudio/android/network/sample/ui/screen/main/list/ProductListAdapter.kt
@@ -6,8 +6,8 @@ import android.widget.TextView
 import androidx.core.view.isVisible
 import me.zhanghai.android.materialprogressbar.MaterialProgressBar
 import ru.surfstudio.android.network.sample.R
-import ru.surfstudio.android.pagination.EasyPaginationAdapter
-import ru.surfstudio.android.pagination.PaginationState
+import ru.surfstudio.android.easyadapter.pagination.EasyPaginationAdapter
+import ru.surfstudio.android.easyadapter.pagination.PaginationState
 
 class ProductListAdapter(onShowMoreListener: () -> Unit) : EasyPaginationAdapter() {