+++ easyadapter/lib-easyadapter-pagination/src/main/java/ru/surfstudio/android/easyadapter/pagination/PaginationState.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package ru.surfstudio.android.pagination;
+package ru.surfstudio.android.easyadapter.pagination;
 
 /**
  * States for {@link EasyPaginationAdapter}.