+++ easyadapter/sample/src/main/java/ru/surfstudio/android/easyadapter/sample/ui/screen/pagination/PaginationableAdapter.kt
@@ -6,8 +6,8 @@ import android.widget.TextView
 import androidx.core.view.isVisible
 import me.zhanghai.android.materialprogressbar.MaterialProgressBar
 import ru.surfstudio.android.easyadapter.sample.R
-import ru.surfstudio.android.pagination.EasyPaginationAdapter
-import ru.surfstudio.android.pagination.PaginationState
+import ru.surfstudio.android.easyadapter.pagination.EasyPaginationAdapter
+import ru.surfstudio.android.easyadapter.pagination.PaginationState
 
 class PaginationableAdapter(onShowMoreListener: () -> Unit) : EasyPaginationAdapter() {