+++ mvi/sample/src/main/java/ru/surfstudio/android/core/mvi/sample/ui/base/middleware/experimental/pagination/PaginationEvent.kt
@@ -1,6 +1,6 @@
 package ru.surfstudio.android.core.mvi.sample.ui.base.middleware.experimental.pagination
 
-import ru.surfstudio.android.pagination.PaginationState
+import ru.surfstudio.android.easyadapter.pagination.PaginationState
 
 /**
  * Событие пагинации