+++ mvp/lib-core-mvp/src/main/java/ru/surfstudio/android/core/mvp/model/LdsPgnScreenModel.java
@@ -16,7 +16,7 @@
 package ru.surfstudio.android.core.mvp.model;
 
 import ru.surfstudio.android.core.mvp.loadstate.LoadStateInterface;
-import ru.surfstudio.android.pagination.PaginationState;
+import ru.surfstudio.android.easyadapter.pagination.PaginationState;
 
 /**
  * модель экрана с поддержкой